requirements.txt
Django==4.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
djangorestframework==3.14.0
djoser==2.1.1
djangorestframework-simplejwt==4.8.0
django-filter==23.1